refactor module structure
