+++ vembrane/cli.py
@@ -1,0 +1,184 @@
+import argparse
+import ast
+import sys
+from collections import defaultdict
+from itertools import islice, chain
+from sys import stderr
+from typing import Iterator
+
+import yaml
+from pysam.libcbcf import VariantFile, VariantHeader, VariantRecord
+
+from . import __version__
+from .errors import InvalidExpression, VembraneError
+from .representations import get_annotation_keys, split_annotation_entry, Environment
+
+
+def main():
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "expression",
+        type=check_filter_expression,
+        help="Filter variants and annotations. If this removes all annotations, "
+        "the variant is removed as well.",
+    )
+    parser.add_argument(
+        "vcf", help="The file containing the variants.", nargs="?", default="-"
+    )
+    parser.add_argument(
+        "--output",
+        "-o",
+        default="-",
+        help="Output file, if not specified, output is written to STDOUT.",
+    )
+    parser.add_argument(
+        "--output-fmt",
+        "-O",
+        default="vcf",
+        choices=["vcf", "bcf", "uncompressed-bcf"],
+        help="Output format.",
+    )
+    parser.add_argument(
+        "--annotation-key",
+        "-k",
+        metavar="FIELDNAME",
+        default="ANN",
+        help="The INFO key for the annotation field.",
+    )
+    parser.add_argument(
+        "--statistics",
+        "-s",
+        metavar="FILE",
+        default=None,
+        help="Write statistics to this file.",
+    )
+    parser.add_argument(
+        "--keep-unmatched",
+        default=False,
+        action="store_true",
+        help="Keep all annotations of a variant if at least one of them passes "
+        "the expression.",
+    )
+    args = parser.parse_args()
+
+    with VariantFile(args.vcf) as vcf:
+        fmt = ""
+        if args.output_fmt == "bcf":
+            fmt = "b"
+        elif args.output_fmt == "uncompressed-bcf":
+            fmt = "u"
+
+        header: VariantHeader = vcf.header
+        header.add_meta("vembraneVersion", __version__)
+        header.add_meta(
+            "vembraneCmd",
+            "vembrane "
+            + " ".join(
+                "'" + arg.replace("'", '"') + '"' if " " in arg else arg
+                for arg in sys.argv[1:]
+            ),
+        )
+
+        records = filter_vcf(
+            vcf,
+            args.expression,
+            args.annotation_key,
+            keep_unmatched=args.keep_unmatched,
+        )
+
+        try:
+            first_record = list(islice(records, 1))
+        except VembraneError as ve:
+            print(ve, file=stderr)
+            exit(1)
+
+        records = chain(first_record, records)
+
+        with VariantFile(args.output, "w" + fmt, header=header,) as out:
+            if args.statistics is not None:
+                records = statistics(records, vcf, args.statistics, args.annotation_key)
+
+            try:
+                for record in records:
+                    out.write(record)
+            except VembraneError as ve:
+                print(ve, file=stderr)
+                exit(1)
+
+
+def filter_vcf(
+    vcf: VariantFile, expression: str, ann_key: str, keep_unmatched: bool = False,
+) -> Iterator[VariantRecord]:
+
+    env = Environment(expression, ann_key, vcf.header)
+
+    record: VariantRecord
+    for idx, record in enumerate(vcf):
+        env.update_from_record(idx, record)
+        if env.filters_annotations():
+            # if the expression contains a reference to the ANN field
+            # get all annotations from the record.info field
+            # (or supply an empty ANN value if the record has no ANN field)
+            try:
+                annotations = record.info[ann_key]
+            except KeyError:
+                annotations = [""]
+            #  … and only keep the annotations where the expression evaluates to true
+            filtered_annotations = [
+                annotation for annotation in annotations if env.evaluate(annotation)
+            ]
+            if not filtered_annotations:
+                # skip this record if filter removed all annotations
+                continue
+            elif not keep_unmatched and (len(annotations) != len(filtered_annotations)):
+                # update annotations if they have actually been filtered
+                record.info[ann_key] = filtered_annotations
+            yield record
+        else:
+            # otherwise, the annotations are irrelevant w.r.t. the expression,
+            # so we can omit them
+            if env.evaluate():
+                yield record
+            else:
+                continue
+
+
+def statistics(
+    records: Iterator[VariantRecord], vcf: VariantFile, filename: str, ann_key: str
+) -> Iterator[VariantRecord]:
+    annotation_keys = get_annotation_keys(vcf.header, ann_key)
+    counter = defaultdict(lambda: defaultdict(lambda: 0))
+    for record in records:
+        for annotation in record.info[ann_key]:
+            for key, raw_value in zip(
+                annotation_keys, split_annotation_entry(annotation)
+            ):
+                value = raw_value.strip()
+                if value:
+                    counter[key][value] += 1
+        yield record
+
+    # reduce dicts with many items, to just one counter
+    for key, subdict in counter.items():
+        if len(subdict) > 10:
+            counter[key] = f"#{len(subdict)}"
+
+    yaml.add_representer(defaultdict, yaml.representer.Representer.represent_dict)
+    with open(filename, "w") as out:
+        yaml.dump(dict(counter), out)
+
+
+def check_filter_expression(expression: str,) -> str:
+    if ".__" in expression:
+        raise InvalidExpression(expression, "The expression must not contain '.__'")
+    try:
+        tree = ast.parse(expression, mode="eval")
+        if isinstance(tree.body, (ast.BoolOp, ast.Compare)):
+            return expression
+        else:
+            # TODO possibly check for ast.Call, func return type
+            return expression
+    except SyntaxError:
+        raise InvalidExpression(
+            expression, "The expression has to be syntactically correct."
+        )
